+++ app/controllers/clients_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class ClientsController < ApplicationController
+  include SessionConcern
+
   before_action :redirect_to_home, only: :new
 
   def new
