+++ app/controllers/planners/homes_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Planners::HomesController < ApplicationController
+  include PlannerConcern
+
   before_action :require_planner_login!
 
   def show
